refactor(LayerMap): tighten preload regression coverage
